+++ src/cpp_audio/synths/QamBeat.cpp
@@ -1,5 +1,5 @@
 #include "QamBeat.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 
 using namespace juce;
