+++ src/cpp_audio/synths/SpatialAngleModulation.cpp
@@ -1,5 +1,5 @@
 #include "SpatialAngleModulation.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 
 using namespace juce;
