+++ src/cpp_audio/synths/Subliminals.cpp
@@ -1,5 +1,5 @@
 #include "Subliminals.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 
 using namespace juce;
